Split out non-required-numbers code
